Week-1/FlavourFuse/backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cohere==3.2.6
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
langchain
loguru==0.6.0
markdown-it-py==2.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==0.4.3
numpy
openai==0.26.4
packaging==23.0
pandas
Pillow==9.4.0
pinecone-client==2.1.0
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rich==13.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.46
streamlit==1.17.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.12.0
Flask-Cors==3.0.10
qdrant-client==1.1.1
git+https://github.com/gaetan1903/py-googletrans@master#egg=googletrans
httpx[http2]